Extracted TimestampType out of Record and Message.
